requirements.txt
flask
msoffcrypto-tool